+++ src/near_connect_static.ts
@@ -1,1 +1,1 @@
-export const NEAR_CONNECT_VERSION = "0.10.0";
+export const NEAR_CONNECT_VERSION = "0.12.0";